Create README.md
Eric and I first created a storyline to set up the user experience (the person who's taking the ticket and parking) versus the "machine doing," which we called behind the scenes. Eric was comfortable working on the function call while I worked on the classes. It was difficult to implement and think about edge cases because the more edge cases we added, the more errors we ran into. Overall, I learned a lot from Eric who helped guide me through the coding.